requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
streamlit==1.28.1
requests==2.31.0
pandas==2.1.3
icalendar==5.0.11
pillow==10.1.0
opencv-python==4.8.1.78
openai>=1.12.0
airportsdata==20241001
timezonefinder==6.2.0
python-multipart==0.0.6
aiofiles==23.2.1